+++ lib/facebook/messenger/server.rb
@@ -144,8 +144,8 @@ module Facebook
       #
       def parsed_body
         @parsed_body ||= JSON.parse(body)
-      rescue JSON::ParserError
-        raise BadRequestError, 'Error parsing request body format'
+      rescue JSON::ParserError => error
+        raise BadRequestError, error
       end
 
       #
@@ -157,15 +157,26 @@ module Facebook
         # Facebook may batch several items in the 'entry' array during
         # periods of high load.
         events['entry'.freeze].each do |entry|
-          # If the application has subscribed to webhooks other than Messenger,
-          # 'messaging' won't be available and it is not relevant to us.
-          next unless entry['messaging'.freeze]
 
           # Facebook may batch several items in the 'messaging' array during
           # periods of high load.
-          entry['messaging'.freeze].each do |messaging|
-            Facebook::Messenger::Bot.receive(messaging)
+          if entry['messaging'.freeze]
+            entry['messaging'.freeze].each do |messaging|
+              Facebook::Messenger::Bot.receive(messaging)
+            end
           end
+          if entry['changes'.freeze]
+            entry['changes'.freeze].each do |change|
+              Facebook::Messenger::Bot.receive(change)
+            end
+          end
+          # d99d
+          if entry['standby'.freeze]
+            entry['standby'.freeze].each do |messaging|
+              Facebook::Messenger::Bot.receive(messaging)
+            end
+          end
+          # .
         end
       end